Revise WAV internals, plus some other stuff. (#45)
* revising wave tech

more work

more work

leader command finally producing accceptable output

nearly there

more worky worky

more

working. needs more testing.

checking

refienments. Doc update

manual update. Implememnt dead zone. Change cocofuse to link aginst fuse-t on macos.

exit when unsupported sample is encountered

cleanup

readme updates

remove symbols and optomize

fix cocofuse for use with fuse-t

adding bit depths

testing 24bit writing

added more tests

more testing

* small cleanup

* change stpcpy. I think it was a typo

* increased delay compensation between blocks

* increased again. Maybe I should make this user configurable.

+++ libcecb/libcecbseek.c
@@ -11,10 +11,9 @@
 #include "cecbpath.h"
 
 
-error_code _cecb_seek(cecb_path_id path, int pos, int mode)
+error_code _cecb_seek(cecb_path_id path, long pos, int mode)
 {
 	error_code ec = 0;
-
 
 	if (path->israw == 1)
 	{
